working on BattleService

+++ src/main/kotlin/net/medrag/theBattle/controller/SquadController.kt
@@ -1,11 +1,11 @@
 package net.medrag.theBattle.controller
 
-import net.medrag.theBattle.model.PLAYER_SESSION
 import net.medrag.theBattle.model.ValidationException
 import net.medrag.theBattle.model.classes.Unitt
 import net.medrag.theBattle.model.dto.UnitDTO
 import net.medrag.theBattle.service.SquadService
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import javax.servlet.http.HttpServletRequest
@@ -22,13 +22,8 @@ class SquadController(@Autowired val squadService: SquadService) {
     //TODO: TODO_SECURITY: requestParam 'name' should be removed in release
     @GetMapping("/getPool")
     fun getPool(@RequestParam(required = false) pName: String?, request: HttpServletRequest): ResponseEntity<List<UnitDTO>> {
-
-        if (pName != null && !pName.isEmpty()) {
-            println("========================================= no session request ================================================")
-            return ResponseEntity.ok(squadService.getPool(pName))
-        }
-
-        val playerName: String = request.getSession(false).getAttribute(PLAYER_SESSION) as String
+        val playerName = extractPlayerName(request, pName)
+        if (playerName.isNullOrBlank()) return ResponseEntity.badRequest().build()
         return ResponseEntity.ok(squadService.getPool(playerName))
     }
 
@@ -40,12 +35,8 @@ class SquadController(@Autowired val squadService: SquadService) {
                request: HttpServletRequest): ResponseEntity<UnitDTO> {
 
         try {
-            if (pName != null && !pName.isEmpty()) {
-                println("========================================= no session request ================================================")
-                return ResponseEntity.ok(squadService.addNewUnit(pName, name, type))
-            }
-
-            val playerName: String = request.getSession(false).getAttribute(PLAYER_SESSION) as String
+            val playerName = extractPlayerName(request, pName)
+            if (playerName.isNullOrBlank()) return ResponseEntity.badRequest().build()
             return ResponseEntity.ok(squadService.addNewUnit(playerName, name, type))
 
         } catch (e: ValidationException) {
@@ -58,14 +49,15 @@ class SquadController(@Autowired val squadService: SquadService) {
     fun delete(@RequestParam unit: Long,
                @RequestParam(required = false) pName: String?,
                request: HttpServletRequest): ResponseEntity<String> {
-        if (pName != null && !pName.isEmpty()) {
-            println("========================================= no session request ================================================")
-            squadService.deleteUnit(unit, pName)
-            return ResponseEntity.ok("RETIRED")
-        }
 
-        val playerName: String = request.getSession(false).getAttribute(PLAYER_SESSION) as String
-        squadService.deleteUnit(unit, playerName)
+        val playerName = extractPlayerName(request, pName)
+        if (playerName.isNullOrBlank())
+            return ResponseEntity("Could not extract player's name", HttpStatus.BAD_REQUEST)
+        try {
+            squadService.deleteUnit(unit, playerName)
+        } catch (e: ValidationException) {
+            return ResponseEntity(e.message, HttpStatus.BAD_REQUEST)
+        }
         return ResponseEntity.ok("RETIRED")
     }
 }
